requirements.txt
paho-mqtt
mysql-connector-python
